wip: start working on ledger simulation service
