+++ manta-pay/src/lib.rs
@@ -27,6 +27,7 @@ extern crate alloc;
 mod test;
 
 pub mod crypto;
+pub mod util;
 
 #[cfg(feature = "groth16")]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "groth16")))]
